Extract chat thread search
