+++ service-worker.js
@@ -179,6 +179,7 @@ const APP_SHELL = [
   // (or PWA install + go-offline) can still open chat / Knowledge Base.
   '/js/chat-images.js',
   '/js/chat-threads.js',
+  '/js/chat-thread-search.js',
   '/js/lens.js',
   '/js/lens-local.js',
   '/js/lens-local-worker.js',
